- Aggiunto valore default a zero nella tabella computo_intervention_rows (voce total); - Aggiunto valore default a zero nella tabella computo_fees_amounts in tutte le colonne; - Quando si aggiunge/modifica un dettaglio (computo) viene creata/aggiornata anche la riga associata dei fees_amount, stessa cosa quando viene eliminato l'ultimo dettaglio, viene eliminata anche la riga in fees_amount; - Creazione tab amount dove viene visualizzata la lista degli interventi inseriti; - Fix vari sul componente Amount; - Aggiunta relazione tra fees_amount e practice; - Fix empty state in price-list.blade.php;

+++ app/Http/Livewire/Practice/Modals/Computo/Tabs/Computo/AddDetails.php
@@ -57,12 +57,10 @@
 				'practice_id'            => $this->practice_id,
 				'intervention_folder_id' => $this->selectedIntervention,
 				'price_row_id'           => $this->row->id,
-				'total'                  => 0
 			]);
-			$this->fees_amount = ComputoFeesAmount::create([
+			$this->fees_amount = ComputoFeesAmount::firstOrCreate([
 				'practice_id'            => $this->practice_id,
 				'intervention_folder_id' => $this->selectedIntervention,
-				'importo_lavori'         => 0,
 			]);
 		}
 
@@ -140,13 +138,19 @@
 					'total' => $this->intervention_row->price_row->price * $this->intervention_row->details->sum('total')
 				]);
 				$this->fees_amount->update([
-					'importo_lavori' => $this->intervention_row->total
+					'importo_lavori' => ComputoInterventionRow::where('practice_id', $this->practice_id)->where('intervention_folder_id', $this->selectedIntervention)->sum('total')
 				]);
 				$this->emit('detail-row-added');
 				$this->closeModal();
 			} else {
 				$this->intervention_row->delete();
-				$this->fees_amount->delete();
+				$total = ComputoInterventionRow::where('practice_id', $this->practice_id)->where('intervention_folder_id', $this->selectedIntervention)->sum('total');
+				$this->fees_amount->update([
+					'importo_lavori' => $total
+				]);
+				if($total === 0) {
+					$this->fees_amount->delete();
+				}
 				$this->emit('detail-row-deleted');
 				$this->closeModal();
 			}
